Update main.py
A work-in-progress physics engine to be used for a semi, yet very - compared to those out there - realistic simulation of molecular life forms and their evolution. Currently, there is W.I.P. energy and matter that can be moved, namely energy distribution as the key physics part, so far. Collision is to be added to the next update of the program.